+++ packages/expo-modules-autolinking/CHANGELOG.md
@@ -13,6 +13,7 @@
 ### 💡 Others
 
 - Check `react-native` and `react-native-tvos` for duplicates in the verify command explicitly ([#40180](https://github.com/expo/expo/pull/40180) by [@kitten](https://github.com/kitten))
+- [Android] Sync flavor dimensions and product flavors from app to expo module ([#40238](https://github.com/expo/expo/pull/40238) by [@kosmydel](https://github.com/kosmydel))
 
 ## 3.0.14 - 2025-10-01